Replaced cmd /c dir with Get-ChildItem again I had originally decided to use cmd /c dir due to performance, but I was finding that there was no way to capture File Path too long errors from dir itself. I replaced it with Get-ChildItem with -erroraction silentlycontinue, so this isn't ideal because it will skip paths where the path or file name is too long and won't report an error. In the future maybe I can dump the directory listing to a file and read from that, since storing the file listing in a variable was taking too much RAM for some trees I was running it on

+++ PermissionsAudit.ps1
@@ -28,7 +28,6 @@ if ((Test-Path -Path "$ErrorFile") -and !(Test-Closed -Path "$ErrorFile")) { thr
 if ((Test-Path -Path "$ReportPrefix.zip") -and !(Test-Closed -Path "$ReportPrefix.zip")) { throw "$ReportPrefix.zip open." }
 
 $PermissionsOutput = @()
-$ErrorOutput = @()
 $FormattedOutput = @()
 
 function Log-PermissionsError ([string]$Path,[string]$Exception) {
@@ -45,18 +44,22 @@ function Get-Permissions {
     )
 
     try {
-        $acl = Get-Acl $Path -ErrorAction Stop
-        Write-Verbose $Path
-        ForEach ($entry in $acl.access) {
-            If ((!$entry.IsInherited -and !$Inheritance) -or ($Inheritance)) { 
-                New-Object -TypeName PSObject -Property @{
-                    Folder = $Path
-                    Access = $entry.FileSystemRights
-                    Control = $entry.AccessControlType
-                    User = $entry.IdentityReference
-                    Inheritance = $entry.IsInherited
+        if (Test-Path -Path "$Path" -ErrorAction Stop) {
+            $acl = Get-Acl "$Path" -ErrorAction Stop
+            Write-Verbose $Path
+            ForEach ($entry in $acl.access) {
+                If ((!$entry.IsInherited -and !$Inheritance) -or ($Inheritance)) { 
+                    New-Object -TypeName PSObject -Property @{
+                        Folder = $Path
+                        Access = $entry.FileSystemRights
+                        Control = $entry.AccessControlType
+                        User = $entry.IdentityReference
+                        Inheritance = $entry.IsInherited
+                    }
                 }
             }
+        } else {
+            throw "Test-Path Failed"
         }
     } catch {
         Log-PermissionsError -Path $Path -Exception $_.exception
@@ -65,34 +68,35 @@ function Get-Permissions {
 
 $PermissionsOutput = $(
     Get-Permissions -Path "$Root" -Inheritance $true;
-    cmd /c dir "$Root" /b /s | %{ Get-Permissions -Path "$_" }
+    Get-ChildItem -Path "$Root" -Recurse -ErrorAction SilentlyContinue | %{ Get-Permissions -Path "$($_.FullName)" }
 )
 
-$PermissionsOutput | Select-Object -Property Folder,Control,Access,Inheritance -Unique | ForEach-Object {
+$FormattedOutput = $PermissionsOutput | Select-Object -Property Folder,Control,Access,Inheritance -Unique | ForEach-Object {
     $CurrentFolder = $_.Folder
     $CurrentControl = $_.Control
     $CurrentAccess = $_.Access
     $CurrentInheritance = $_.Inheritance
-    $CurrentUsers = ""
-    $PermissionsOutput | Where-Object { $_.Folder -eq $CurrentFolder -and $_.Control -eq $CurrentControl -and $_.Access -eq $CurrentAccess -and $_.Inheritance -eq $CurrentInheritance } | ForEach-Object {
-        $CurrentUsers = "$CurrentUsers" + $_.User + ";"
-    }
-    $FormattedOutput += New-Object -TypeName PSObject -Property @{
+    $CurrentUsers = $($PermissionsOutput | 
+        Where-Object { $_.Folder -eq $CurrentFolder -and $_.Control -eq $CurrentControl -and $_.Access -eq $CurrentAccess -and $_.Inheritance -eq $CurrentInheritance } | 
+        Select -ExpandProperty User) -join ";"
+
+    New-Object -TypeName PSObject -Property @{
         Folder = $CurrentFolder
         Access = $CurrentAccess
         Control = $CurrentControl
         Users = $CurrentUsers
         Inheritance = $CurrentInheritance
     }
-}
-
-$FormattedOutput | Select-Object -Property Folder,Users,Control,Access,Inheritance | Export-Csv -NoTypeInformation -Path $OutputFile
+} | Select-Object -Property Folder,Users,Control,Access,Inheritance | Export-Csv -NoTypeInformation -Path $OutputFile
 
 $ReportFiles = @()
 if (Test-Path -Path "$OutputFile") { $ReportFiles += $OutputFile }
 if (Test-Path -Path "$ErrorFile") { $ReportFiles += $ErrorFile }
+
 Compress-Archive -LiteralPath $ReportFiles -DestinationPath "$ReportPrefix.zip" -Force
-if (Test-Path -Path "$OutputFile") { Remove-Item -Path "$OutputFile" -Force }
-if (Test-Path -Path "$ErrorFile") { Remove-Item -Path "$ErrorFile" -Force }
+if (Test-Path -Path "$ReportPrefix.zip") {
+    if (Test-Path -Path "$OutputFile") { Remove-Item -Path "$OutputFile" -Force }
+    if (Test-Path -Path "$ErrorFile") { Remove-Item -Path "$ErrorFile" -Force }
+}
 
 Write-Output "$ReportPrefix.zip"
